requirements.txt
numpy
pandas
scipy
lightfm
nmslib-metabrainz-2.1.3
plotly